Tagged hash trees, take 1
This first step adds tagging to all the internal
machinery of a hash tree, but does not change
the API.

Internal representation of a bucket is changed
from 

	[{SubID, <<hash>>}] 

to
	[{SubID, [{tag, <<hash>>}]}]; 

i.e. all the inner nodes have an orddict of 
tag->hash for that specific tag of the hash tree.

A new option 

	{tag_fun, fun((binary()) -> [atom()])}

controls the tagging.  The argument is the Key
as given to the hash tree, which is typically
term_to_binary({Bucket,Key}), allowing the 
client to decode the BK and determine tags from
the real {K,V}.

We will then expose access to tagged variants
of the API only as needed by new functionality.